+++ CriticalMass/IDProvider.swift
@@ -9,4 +9,7 @@ import Foundation
 
 public protocol IDProvider {
     var id: String { get }
+    var token: String { get }
+    
+    static func hash(id: String, currentDate: Date) -> String
 }