+++ CriticalMass/Logger.swift
@@ -9,10 +9,25 @@
 import Foundation
 import os.log
 
+extension OSLog {
+    private static var subsystem = Bundle.main.bundleIdentifier!
+
+    static let viewManagement = OSLog(subsystem: subsystem, category: "viewManagement")
+}
+
 class Logger {
     static func log(_ type: OSLogType, log: OSLog, _ message: StaticString) {
         if #available(iOS 12.0, macOS 10.14, *) {
             os_log(type, log: log, message)
+        } else {
+            print(message)
+        }
+    }
+
+    static func log(_ type: OSLogType, log: OSLog, _ message: StaticString, parameter: String) {
+        if #available(iOS 12.0, *) {
+            let logMessage = "\(message): \(parameter)"
+            os_log(type, log: log, "%@", logMessage)
         } else {
             print(message)
         }