+++ CriticalMass/TwitterViewController.swift
@@ -7,9 +7,21 @@
 
 import UIKit
 
-class TwitterViewController: UIViewController {
+class TwitterViewController: UIViewController, ContentStatePresentable {
     private let messagesTableViewController = MessagesTableViewController<TweetTableViewCell>(style: .plain)
     private let twitterManager: TwitterManager
+
+    var contentStateViewController: UIViewController? {
+        willSet {
+            if let controller = contentStateViewController {
+                controller.remove()
+            }
+            guard let viewController = newValue else {
+                return
+            }
+            addAndLayoutStateView(viewController, in: view)
+        }
+    }
 
     init(twitterManager: TwitterManager) {
         self.twitterManager = twitterManager
@@ -22,33 +34,34 @@ class TwitterViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
         configureMessagesTableViewController()
+        twitterManager.getTweets()
     }
 
     private func configureMessagesTableViewController() {
         messagesTableViewController.noContentMessage = String.twitterNoData
-        messagesTableViewController.messages = twitterManager.getTweets()
         messagesTableViewController.pullToRefreshTrigger = twitterManager.loadTweets
         messagesTableViewController.selectMessageTrigger = { [weak self] selectedTweet in
             self?.openTweet(selectedTweet)
         }
 
-        twitterManager.updateTweetsCallback = { [weak self] tweets in
-            self?.messagesTableViewController.update(messages: tweets)
+        twitterManager.updateContentStateCallback = { [weak self] contentState in
+            switch contentState {
+            case let .results(tweets):
+                self?.messagesTableViewController.update(messages: tweets)
+                self?.contentStateViewController = nil
+            case let .loading(loadingViewController):
+                self?.contentStateViewController = loadingViewController
+            case let .error(errorStateViewController):
+                errorStateViewController.reloadHandler = { [weak self] in
+                    self?.twitterManager.getTweets()
+                }
+                self?.contentStateViewController = errorStateViewController
+            }
         }
 
-        addChild(messagesTableViewController)
-        view.addSubview(messagesTableViewController.view)
-        messagesTableViewController.didMove(toParent: self)
-        messagesTableViewController.view.translatesAutoresizingMaskIntoConstraints = false
-
-        view.addConstraints([
-            NSLayoutConstraint(item: messagesTableViewController.view!, attribute: .top, relatedBy: .equal, toItem: view, attribute: .topMargin, multiplier: 1, constant: 0),
-            NSLayoutConstraint(item: messagesTableViewController.view!, attribute: .width, relatedBy: .equal, toItem: view, attribute: .width, multiplier: 1, constant: 0),
-            NSLayoutConstraint(item: messagesTableViewController.view!, attribute: .centerX, relatedBy: .equal, toItem: view, attribute: .centerX, multiplier: 1, constant: 0),
-            NSLayoutConstraint(item: messagesTableViewController.view!, attribute: .bottom, relatedBy: .equal, toItem: view, attribute: .bottom, multiplier: 1, constant: 0),
-        ])
+        add(messagesTableViewController)
+        layout(messagesTableViewController)
         // inset tableView seperator
         messagesTableViewController.tableView.separatorInset = UIEdgeInsets(top: 0.0, left: 73.0, bottom: 0.0, right: 0.0)
     }