+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Changelog for Critical Maps iOS
 
 - Fix: Don't update content when slightly swipe down in Social Modal
 - Fix whitespace only chat message can not be send anymore.
+- Fix dark mode for Rules Detail
+
+### Added
+
+- Loading and ErrorStateView for Twitter section
 
 ## [3.4.0] - 2019-10-08